Move Testing Functions to testing.go file
