+++ cli/consignment/client_test.go
@@ -1,30 +1,8 @@
 package main
 
 import (
-	"fmt"
-	"path/filepath"
-	"reflect"
-	"runtime"
 	"testing"
 )
-
-// ok fails the test if an err is not nil.
-func ok(tb testing.TB, err error) {
-	if err != nil {
-		_, file, line, _ := runtime.Caller(1)
-		fmt.Printf("\033[31m%s:%d: unexpected error: %s\033[39m\n\n", filepath.Base(file), line, err.Error())
-		tb.FailNow()
-	}
-}
-
-// equals fails the test if exp is not equal to act.
-func equals(tb testing.TB, exp, act interface{}) {
-	if !reflect.DeepEqual(exp, act) {
-		_, file, line, _ := runtime.Caller(1)
-		fmt.Printf("\033[31m%s:%d:\n\n\texp: %#v\n\n\tgot: %#v\033[39m\n\n", filepath.Base(file), line, exp, act)
-		tb.FailNow()
-	}
-}
 
 func TestParseFile(t *testing.T) {
 	c, err := parseFile(defaultFilename)
@@ -43,9 +21,9 @@ func TestParseFile(t *testing.T) {
 		{fieldName: "Containers", got: len(c.GetContainers()), want: 3},
 	}
 
-	for _, tt := range testCases {
-		t.Run(tt.fieldName, func(t *testing.T) {
-			equals(t, tt.got, tt.want)
+	for _, tc := range testCases {
+		t.Run(tc.fieldName, func(t *testing.T) {
+			equals(t, tc.got, tc.want)
 		})
 	}
 }
